Move User type together with remaining types
