--- src/types/User.ts
@@ -1,5 +1,0 @@
-export type User = {
-  id: string;
-  name: string;
-  isOnline: boolean;
-};
